final-pjt-back/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.57.0
django-appconf==1.0.5
django-cors-headers==4.3.1
django-environ==0.11.2
django-imagekit==5.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
oauthlib==3.2.2
pilkit==3.0
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
